requirements.txt
sphinx
sphinx-rtd-theme
urllib3

scipy>=1.9
numpy
cvxpy
matplotlib
pyinterval
sympy
